requirements.txt
# Core ML/DL Framework
torch
torchvision

# Hugging Face ecosystem
datasets
# pin to a stable version that works with our training scripts
transformers>=4.41.0
accelerate>=0.33.0
sentencepiece
huggingface_hub
# diffusers for PixArt-Sigma
diffusers==0.30.0
safetensors

# Scientific computing and Image Processing
numpy
scipy
Pillow

# Dependencies for Infinity Model
einops
omegaconf
timm

peft
# High-performance attention (optional; leave unpinned)
# xformers
# flash-attn